Remove network from ContractInstance ID
Since contract instances may now not have a transaction hash, and
consequently, may be associated with a Network whose historicBlock is
not necessarily the block in which the instance was created.

As a result, the `network` field now lacks determinism, so other fields
must be used for ID: switch to use `contract`, `address`, and
`creation`.